chore: Register domain
